+++ src/utils/render.js
@@ -21,20 +21,4 @@ export const createElement = (template) => {
   return newElement.firstChild;
 };
 
-export const replace = (newComponent, oldComponent) => {
-  const parentElement = oldComponent.getElement().parentElement;
-  const newElement = newComponent.getElement();
-  const oldElement = oldComponent.getElement();
-
-  const isExistElements = !!(parentElement && newElement && oldElement);
-
-  if (isExistElements && parentElement.contains(oldElement)) {
-    parentElement.replaceChild(newElement, oldElement);
-  }
-};
-
-export const remove = (component) => {
-  component.getElement().remove();
-  component.removeElement();
-};
